step_1: implementation still in .h files, dont understand error: "undefined reference .." must be someting in LivingRoom.h
